Refactor tests set up flow
